Send Events when comms are turned off or repaired
To allow for this, a temporary design choice has been taken, requiring
us to add a reference to the Game and to the eventManager to each
subsystem.

If we decide to keep this design, we should at least move them to the
ISystemType class, although I think I'd rather store a reference to
the InnerShipStatus that created it and have it send the Events. Long
Term it could allow us to support the same object corresponding to
different sabotages depending on the map.

Another possibility would be to pass a callback at creation of the
system that sends the correct event.