Improve cross-compatibility
Official servers are allowing multiple game versions to crossplay.
Impostor currently has a binary toggle that allows or forbids *all*
crossplay. This PR adds the so-called CompatibiltyManager which puts all
supported network versions in groups: if two versions are in the same
group, they are allowed to play together.

This feature is also exposed as API so that plugins can mess this up
on runtime without requiring server restarts every time Innersloth
releases a small patch which bumps the network version but does not
change the network protocol enough.

Also bump the minor version because we're introducing API and changing
behaviour compared to previous versions.