Let hosts handle exiles
Previously Impostor liked to kill exiled players server side to raise an
event, however as PlayerInfo is now networked, it also killed players
client side. This caused meetings to immediately conclude if the final
impostor was voted out, ruining the cutscene.

So wait for hosts to actually kill the player