Fix regression in game started detection
In Impostor 1.9.x games only started when all players had a role set.
Currently games will start when the first role is assigned, with all
other players defaulting to Crewmate, ignoring the other SetRole
commands that Impostor has not yet processed.

This commit reverts to the previous correct behaviour where SetRole
needs to be called for all players. Also ignore RoleType when
deserializing, as Vanilla roles are not nullable